requirements.txt
numpy==1.26.4
pandas==2.2.2
opencv-python==4.8.1.78
torch
torchvision
tqdm
Pillow
gdown
ultralytics
insightface
onnxruntime-gpu==1.19.2